+++ src/PVOutput.Net/Objects/Modules/Readers/BatchStatusPostResultStringReader.cs
@@ -12,7 +12,7 @@ namespace PVOutput.Net.Objects.Modules.Readers
             var properties = new Action<IBatchStatusPostResult, string>[]
             {
                 (t, s) => t.Timestamp = FormatHelper.ParseDate(s),
-                (t, s) => t.Timestamp = s.Equals("NaN", StringComparison.OrdinalIgnoreCase) ? t.Timestamp : t.Timestamp.Add(FormatHelper.ParseTime(s).TimeOfDay),
+                (t, s) => t.Timestamp = s.Equals("NaN", StringComparison.OrdinalIgnoreCase) ? t.Timestamp : t.Timestamp.Add(FormatHelper.ParseTime(s)),
                 (t, s) => t.AddedOrUpdated = FormatHelper.GetValueOrDefault<int>(s) == 1
             };