+++ src/PVOutput.Net/Objects/Modules/Readers/OutputObjectStringReader.cs
@@ -17,7 +17,7 @@ namespace PVOutput.Net.Objects.Modules.Readers
                 (t, s) => t.EnergyExported = FormatHelper.GetValueOrDefault<int>(s),
                 (t, s) => t.EnergyUsed = FormatHelper.GetValueOrDefault<int>(s),
                 (t, s) => t.PeakPower = FormatHelper.GetValue<int>(s),
-                (t, s) => t.PeakTime = s.Equals("NaN", StringComparison.OrdinalIgnoreCase) ? (DateTime?)null : t.OutputDate.Add(FormatHelper.ParseTime(s).TimeOfDay),
+                (t, s) => t.PeakTime = s.Equals("NaN", StringComparison.OrdinalIgnoreCase) ? (TimeSpan?)null : FormatHelper.ParseTime(s),
                 (t, s) => t.Condition = FormatHelper.DescriptionToEnumValue<WeatherCondition>(s),
                 (t, s) => t.MinimumTemperature = FormatHelper.GetValue<int>(s),
                 (t, s) => t.MaximumTemperature = FormatHelper.GetValue<int>(s),