+++ tests/PVOutput.Net.Tests/Modules/Output/OutputBuilderTests.cs
@@ -76,22 +76,9 @@ namespace PVOutput.Net.Tests.Modules.Output
         public void OutputPostBuilder_WithPeakTime_SetsPeakTime()
         {
             var builder = new OutputPostBuilder().SetDate(new DateTime(2020, 1, 1))
-                .SetPeakTime(new DateTime(2020, 1, 1, 10, 10, 0));
+                .SetPeakTime(new TimeSpan(10, 10, 0));
 
-            Assert.AreEqual(new DateTime(2020, 1, 1, 10, 10, 0), builder.OutputPost.PeakTime.Value);
-        }
-
-
-        [Test]
-        public void OutputPostBuilder_WithPeakTimeOnOtherDate_CannotBuild()
-        {
-            var builder = new OutputPostBuilder().SetDate(DateTime.Today)
-                .SetPeakTime(new DateTime(2020, 1, 1, 10, 10, 0));
-
-            Assert.Throws<InvalidOperationException>(() =>
-            {
-                builder.Build();
-            });
+            Assert.AreEqual(new TimeSpan(10, 10, 0), builder.OutputPost.PeakTime.Value);
         }
 
         [Test]